requirements.txt
pillow==8.2.0
django==3.2.4
django-crispy-forms==1.12.0
django-cleanup==5.2.0
django-ckeditor==6.1.0
flake8==3.9.2